+++ crates/barracuda/src/shaders/math/matmul_cpu_tiled_f64.wgsl
@@ -1,10 +1,10 @@
-// matmul_cpu_tiled.wgsl — Double-buffered 32×32 tiled matmul for CPU (llvmpipe)
+// matmul_cpu_tiled_f64.wgsl — Double-buffered 32×32 tiled matmul for CPU (f64 canonical)
 //
 // Absorbed from neuralSpring local evolutions (neuralSpring handoff #11).
 //
 // Key optimisations vs the naive 16×16 tiled shader:
 //   1. TILE = 32 → 4× more arithmetic reuse per global-memory load;
-//      two 32×32 f32 tiles ≈ 8 KB, fits in typical L1 data cache.
+//      two 32×32 f64 tiles ≈ 16 KB, fits in typical L1 data cache.
 //   2. Double-buffered tiles — while computing tileX_A / tileX_B, the
 //      same invocation preloads the NEXT tile into tileY_A / tileY_B.
 //      On llvmpipe this keeps the LLVM vectoriser fed without stalls.
@@ -26,17 +26,16 @@ struct MatMulParams {
     _padding: u32,
 }
 
-@group(0) @binding(0) var<storage, read>       A:      array<f32>;
-@group(0) @binding(1) var<storage, read>       B:      array<f32>;
-@group(0) @binding(2) var<storage, read_write> C:      array<f32>;
+@group(0) @binding(0) var<storage, read>       A:      array<f64>;
+@group(0) @binding(1) var<storage, read>       B:      array<f64>;
+@group(0) @binding(2) var<storage, read_write> C:      array<f64>;
 @group(0) @binding(3) var<uniform>             params: MatMulParams;
 
 // Two pairs of shared-memory tiles for double buffering.
-// Each tile holds 32×32 f32 = 4 KB.  Total workgroup memory: 4 × 4 KB = 16 KB.
-var<workgroup> tileA_curr: array<f32, 1024>;   // current A tile
-var<workgroup> tileB_curr: array<f32, 1024>;   // current B tile
-var<workgroup> tileA_next: array<f32, 1024>;   // prefetch A tile
-var<workgroup> tileB_next: array<f32, 1024>;   // prefetch B tile
+var<workgroup> tileA_curr: array<f64, 1024>;   // current A tile
+var<workgroup> tileB_curr: array<f64, 1024>;   // current B tile
+var<workgroup> tileA_next: array<f64, 1024>;   // prefetch A tile
+var<workgroup> tileB_next: array<f64, 1024>;   // prefetch B tile
 
 // 16×16 workgroup, each thread computes a 2×2 output block → 32×32 output tile.
 @compute @workgroup_size(16, 16)
@@ -53,36 +52,33 @@ fn main(
     let lcol = local_id.x;
 
     // 4-element accumulators for the 2×2 output block:
-    //   acc_rr[r][c]  where r,c ∈ {0,1}
-    var acc00 = 0.0f;
-    var acc01 = 0.0f;
-    var acc10 = 0.0f;
-    var acc11 = 0.0f;
+    var acc00 = 0.0;
+    var acc01 = 0.0;
+    var acc10 = 0.0;
+    var acc11 = 0.0;
 
     let num_tiles = (params.k + TILE - 1u) / TILE;
 
     // ── Pre-load tile 0 into "current" buffers ────────────────────────────
-    // Each 16×16 workgroup cooperatively loads the 32×32 tile:
-    // every invocation loads 4 elements (2 rows × 2 cols).
     for (var dr = 0u; dr < 2u; dr = dr + 1u) {
         for (var dc = 0u; dc < 2u; dc = dc + 1u) {
             let srow = lrow * 2u + dr;
             let scol = lcol * 2u + dc;
 
             let a_row = workgroup_id.y * TILE + srow;
-            let a_col = scol;   // tile 0: col offset = 0
+            let a_col = scol;
             if (a_row < params.m && a_col < params.k) {
                 tileA_curr[srow * TILE + scol] = A[a_row * params.k + a_col];
             } else {
-                tileA_curr[srow * TILE + scol] = 0.0f;
+                tileA_curr[srow * TILE + scol] = 0.0;
             }
 
-            let b_row = srow;   // tile 0: row offset = 0
+            let b_row = srow;
             let b_col = workgroup_id.x * TILE + scol;
             if (b_row < params.k && b_col < params.n) {
                 tileB_curr[srow * TILE + scol] = B[b_row * params.n + b_col];
             } else {
-                tileB_curr[srow * TILE + scol] = 0.0f;
+                tileB_curr[srow * TILE + scol] = 0.0;
             }
         }
     }
@@ -91,8 +87,6 @@ fn main(
     // ── Double-buffered tile loop ─────────────────────────────────────────
     for (var t = 0u; t < num_tiles; t = t + 1u) {
 
-        // Prefetch next tile (if it exists) into "next" buffers while we
-        // compute with "current".  On CPU this keeps the load pipeline busy.
         let next_t = t + 1u;
         if (next_t < num_tiles) {
             for (var dr = 0u; dr < 2u; dr = dr + 1u) {
@@ -105,7 +99,7 @@ fn main(
                     if (a_row < params.m && a_col < params.k) {
                         tileA_next[srow * TILE + scol] = A[a_row * params.k + a_col];
                     } else {
-                        tileA_next[srow * TILE + scol] = 0.0f;
+                        tileA_next[srow * TILE + scol] = 0.0;
                     }
 
                     let b_row = next_t * TILE + srow;
@@ -113,19 +107,16 @@ fn main(
                     if (b_row < params.k && b_col < params.n) {
                         tileB_next[srow * TILE + scol] = B[b_row * params.n + b_col];
                     } else {
-                        tileB_next[srow * TILE + scol] = 0.0f;
+                        tileB_next[srow * TILE + scol] = 0.0;
                     }
                 }
             }
         }
 
         // ── Compute 2×2 block from current tile (4× k-unrolled) ──────────
-        // Unrolling by 4 produces 4 independent FMA chains — each chain can
-        // be scheduled independently by llvmpipe's ILP scheduler.
         let k_limit = min(TILE, params.k - t * TILE);
         var k = 0u;
 
-        // Unrolled body: 4 steps at a time
         for (; k + 4u <= k_limit; k = k + 4u) {
             let a0_r0 = tileA_curr[lrow * 2u       * TILE + k];
             let a1_r0 = tileA_curr[lrow * 2u       * TILE + k + 1u];
@@ -145,13 +136,11 @@ fn main(
             let b2_c1 = tileB_curr[(k + 2u) * TILE + lcol * 2u + 1u];
             let b3_c1 = tileB_curr[(k + 3u) * TILE + lcol * 2u + 1u];
 
-            // fma() → LLVM fmuladd → vfmadd (AVX-512 on x86)
             acc00 = fma(a0_r0, b0_c0, fma(a1_r0, b1_c0, fma(a2_r0, b2_c0, fma(a3_r0, b3_c0, acc00))));
             acc01 = fma(a0_r0, b0_c1, fma(a1_r0, b1_c1, fma(a2_r0, b2_c1, fma(a3_r0, b3_c1, acc01))));
             acc10 = fma(a0_r1, b0_c0, fma(a1_r1, b1_c0, fma(a2_r1, b2_c0, fma(a3_r1, b3_c0, acc10))));
             acc11 = fma(a0_r1, b0_c1, fma(a1_r1, b1_c1, fma(a2_r1, b2_c1, fma(a3_r1, b3_c1, acc11))));
         }
-        // Tail (< 4 remaining k-steps)
         for (; k < k_limit; k = k + 1u) {
             let a_r0 = tileA_curr[lrow * 2u        * TILE + k];
             let a_r1 = tileA_curr[(lrow * 2u + 1u) * TILE + k];
@@ -163,11 +152,7 @@ fn main(
             acc11 = fma(a_r1, b_c1, acc11);
         }
 
-        // Swap current ↔ next: barrier ensures next is fully loaded.
         workgroupBarrier();
-        // Copy next → current for the following iteration.
-        // (WGSL has no pointer aliasing, so we do an explicit copy via
-        // the same 2×2 load pattern each thread already owns.)
         if (next_t < num_tiles) {
             for (var dr = 0u; dr < 2u; dr = dr + 1u) {
                 for (var dc = 0u; dc < 2u; dc = dc + 1u) {
