+++ crates/barracuda/src/shaders/math/matmul_gpu_evolved_f64.wgsl
@@ -1,18 +1,7 @@
-// matmul_gpu_evolved.wgsl — Double-buffered 32×32 tiled matmul for large GPU matrices
+// matmul_gpu_evolved_f64.wgsl — Double-buffered 32×32 tiled matmul for large GPU matrices (f64 canonical)
 //
 // Absorbed from neuralSpring local evolutions (neuralSpring handoff #11).
 // Effective when M ≥ 256 and N ≥ 256 on a discrete/integrated GPU.
-//
-// Key optimisations vs matmul_tiled.wgsl (16×16 single-buffer):
-//   1. TILE = 32 → 4× more arithmetic reuse per global-memory round-trip;
-//      reduces bandwidth pressure which is the dominant cost on large matmuls.
-//   2. Double-buffered tiles — issues loads for tile t+1 while the ALU is
-//      computing tile t; overlaps memory latency with arithmetic (SM-style
-//      warp scheduling on NVIDIA / wave64 scheduling on AMD).
-//   3. 2×2 micro-kernel — each thread accumulates 4 outputs, doubling
-//      arithmetic intensity (FLOPs per byte loaded) vs the single-element kernel.
-//   4. 4× k-loop unroll — creates 4 independent FMA dependency chains,
-//      filling SM warp latency windows (8 cycles/DFMA on Turing).
 //
 // C = A × B,  A:[M,K], B:[K,N], C:[M,N]
 
@@ -25,18 +14,16 @@ struct MatMulParams {
     _padding: u32,
 }
 
-@group(0) @binding(0) var<storage, read>       A:      array<f32>;
-@group(0) @binding(1) var<storage, read>       B:      array<f32>;
-@group(0) @binding(2) var<storage, read_write> C:      array<f32>;
+@group(0) @binding(0) var<storage, read>       A:      array<f64>;
+@group(0) @binding(1) var<storage, read>       B:      array<f64>;
+@group(0) @binding(2) var<storage, read_write> C:      array<f64>;
 @group(0) @binding(3) var<uniform>             params: MatMulParams;
 
-// Double-buffered 32×32 tiles (4 KB each, 16 KB total workgroup memory).
-var<workgroup> tileA_curr: array<f32, 1024>;
-var<workgroup> tileB_curr: array<f32, 1024>;
-var<workgroup> tileA_next: array<f32, 1024>;
-var<workgroup> tileB_next: array<f32, 1024>;
+var<workgroup> tileA_curr: array<f64, 1024>;
+var<workgroup> tileB_curr: array<f64, 1024>;
+var<workgroup> tileA_next: array<f64, 1024>;
+var<workgroup> tileB_next: array<f64, 1024>;
 
-// 16×16 workgroup; each thread handles a 2×2 output block → 32×32 output tile/WG.
 @compute @workgroup_size(16, 16)
 fn main(
     @builtin(global_invocation_id) global_id:    vec3<u32>,
@@ -49,10 +36,10 @@ fn main(
     let lrow = local_id.y;
     let lcol = local_id.x;
 
-    var acc00 = 0.0f;
-    var acc01 = 0.0f;
-    var acc10 = 0.0f;
-    var acc11 = 0.0f;
+    var acc00 = 0.0;
+    var acc01 = 0.0;
+    var acc10 = 0.0;
+    var acc11 = 0.0;
 
     let num_tiles = (params.k + TILE - 1u) / TILE;
 
@@ -67,7 +54,7 @@ fn main(
             if (a_row < params.m && a_col < params.k) {
                 tileA_curr[srow * TILE + scol] = A[a_row * params.k + a_col];
             } else {
-                tileA_curr[srow * TILE + scol] = 0.0f;
+                tileA_curr[srow * TILE + scol] = 0.0;
             }
 
             let b_row = srow;
@@ -75,7 +62,7 @@ fn main(
             if (b_row < params.k && b_col < params.n) {
                 tileB_curr[srow * TILE + scol] = B[b_row * params.n + b_col];
             } else {
-                tileB_curr[srow * TILE + scol] = 0.0f;
+                tileB_curr[srow * TILE + scol] = 0.0;
             }
         }
     }
@@ -84,8 +71,6 @@ fn main(
     // ── Double-buffered tile loop ─────────────────────────────────────────
     for (var t = 0u; t < num_tiles; t = t + 1u) {
 
-        // Prefetch tile t+1 while computing tile t.
-        // On GPU, this overlaps async global-memory loads with the FMA pipeline.
         let next_t = t + 1u;
         if (next_t < num_tiles) {
             for (var dr = 0u; dr < 2u; dr = dr + 1u) {
@@ -98,7 +83,7 @@ fn main(
                     if (a_row < params.m && a_col < params.k) {
                         tileA_next[srow * TILE + scol] = A[a_row * params.k + a_col];
                     } else {
-                        tileA_next[srow * TILE + scol] = 0.0f;
+                        tileA_next[srow * TILE + scol] = 0.0;
                     }
 
                     let b_row = next_t * TILE + srow;
@@ -106,19 +91,17 @@ fn main(
                     if (b_row < params.k && b_col < params.n) {
                         tileB_next[srow * TILE + scol] = B[b_row * params.n + b_col];
                     } else {
-                        tileB_next[srow * TILE + scol] = 0.0f;
+                        tileB_next[srow * TILE + scol] = 0.0;
                     }
                 }
             }
         }
 
         // ── 2×2 micro-kernel, 4× k-unrolled ──────────────────────────────
-        // Four independent accumulator chains hide FMA issue latency.
         let k_limit = min(TILE, params.k - t * TILE);
         var k = 0u;
 
         for (; k + 4u <= k_limit; k = k + 4u) {
-            // Load 2 A-rows × 4 k-steps
             let a00 = tileA_curr[lrow * 2u        * TILE + k];
             let a01 = tileA_curr[lrow * 2u        * TILE + k + 1u];
             let a02 = tileA_curr[lrow * 2u        * TILE + k + 2u];
@@ -128,7 +111,6 @@ fn main(
             let a12 = tileA_curr[(lrow * 2u + 1u) * TILE + k + 2u];
             let a13 = tileA_curr[(lrow * 2u + 1u) * TILE + k + 3u];
 
-            // Load 2 B-cols × 4 k-steps
             let b00 = tileB_curr[k        * TILE + lcol * 2u];
             let b10 = tileB_curr[(k + 1u) * TILE + lcol * 2u];
             let b20 = tileB_curr[(k + 2u) * TILE + lcol * 2u];
@@ -138,13 +120,11 @@ fn main(
             let b21 = tileB_curr[(k + 2u) * TILE + lcol * 2u + 1u];
             let b31 = tileB_curr[(k + 3u) * TILE + lcol * 2u + 1u];
 
-            // 4 independent FMA chains → 4-way ILP on GPU issue units
             acc00 = fma(a00, b00, fma(a01, b10, fma(a02, b20, fma(a03, b30, acc00))));
             acc01 = fma(a00, b01, fma(a01, b11, fma(a02, b21, fma(a03, b31, acc01))));
             acc10 = fma(a10, b00, fma(a11, b10, fma(a12, b20, fma(a13, b30, acc10))));
             acc11 = fma(a10, b01, fma(a11, b11, fma(a12, b21, fma(a13, b31, acc11))));
         }
-        // Tail: remaining < 4 k-steps
         for (; k < k_limit; k = k + 1u) {
             let a_r0 = tileA_curr[lrow * 2u        * TILE + k];
             let a_r1 = tileA_curr[(lrow * 2u + 1u) * TILE + k];
@@ -156,7 +136,6 @@ fn main(
             acc11 = fma(a_r1, b_c1, acc11);
         }
 
-        // Synchronise then swap current ↔ prefetched buffers.
         workgroupBarrier();
         if (next_t < num_tiles) {
             for (var dr = 0u; dr < 2u; dr = dr + 1u) {
