+++ crates/barracuda/src/shaders/math/matrix_power_f64.wgsl
@@ -1,4 +1,4 @@
-// Matrix Power - M^n via exponentiation by squaring (complete implementation)
+// Matrix Power - M^n via exponentiation by squaring (f64 canonical)
 // Efficiently computes matrix raised to integer power
 //
 // Algorithm: Exponentiation by squaring
@@ -17,16 +17,16 @@ struct Params {
 }
 
 @group(0) @binding(0) var<uniform> params: Params;
-@group(0) @binding(1) var<storage, read> matrix_a: array<f32>;  // Left matrix
-@group(0) @binding(2) var<storage, read> matrix_b: array<f32>;  // Right matrix
-@group(0) @binding(3) var<storage, read_write> output: array<f32>;  // Result
+@group(0) @binding(1) var<storage, read> matrix_a: array<f64>;  // Left matrix
+@group(0) @binding(2) var<storage, read> matrix_b: array<f64>;  // Right matrix
+@group(0) @binding(3) var<storage, read_write> output: array<f64>;  // Result
 
 // Matrix multiplication (used iteratively for power computation)
 @compute @workgroup_size(16, 16, 1)
 fn matrix_multiply(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let row = global_id.y;
     let col = global_id.x;
-    
+
     if (row >= params.size || col >= params.size) {
         return;
     }
@@ -35,7 +35,7 @@ fn matrix_multiply(@builtin(global_invocation_id) global_id: vec3<u32>) {
     for (var k = 0u; k < params.size; k++) {
         sum += matrix_a[row * params.size + k] * matrix_b[k * params.size + col];
     }
-    
+
     output[row * params.size + col] = sum;
 }
 
@@ -44,7 +44,7 @@ fn matrix_multiply(@builtin(global_invocation_id) global_id: vec3<u32>) {
 fn init_identity(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let row = global_id.y;
     let col = global_id.x;
-    
+
     if (row >= params.size || col >= params.size) {
         return;
     }
